LPD-63103 BufferedIncrement requires increment to be the last argument which was why it was never working
